feat 单位问题

+++ src/main/java/com/base/sbc/module/material/controller/MaterialController.java
@@ -424,8 +424,8 @@ public class MaterialController extends BaseController {
     public ApiResult spaceUsage(){
         String userId = userUtils.getUserCompany().getUserId();
         MaterialSpaceInfoVo materialSpaceInfoVo = new MaterialSpaceInfoVo();
-        materialSpaceInfoVo.setUsageSpaceSize(materialService.getFileSize(userId, null));
-        Long personSpace = storageSpacePersonService.getPersonSpace(userId, "1");
+        materialSpaceInfoVo.setUsageSpaceSize(Double.valueOf(materialService.getFileSize(userId, null)));
+        Double personSpace = storageSpacePersonService.getPersonSpace(userId, "1");
         materialSpaceInfoVo.setTotalSpaceSize(personSpace*1073741824);
         return ApiResult.success("更新成功",materialSpaceInfoVo);
     }
