+++ src/main/java/com/base/sbc/module/storageSpace/service/impl/StorageSpacePersonServiceImpl.java
@@ -134,7 +134,7 @@ public class StorageSpacePersonServiceImpl extends BaseServiceImpl<StorageSpaceP
         if ( StringUtils.isEmpty(spacePerson.getOwnerSpace())){
             throw new OtherException("个人可用空间已被清理，请重新分配空间");
         }
-        long ownSpacer = StringUtils.isEmpty(spacePerson.getOwnerSpace()) ? 0L : Long.parseLong(spacePerson.getOwnerSpace());
+        double ownSpacer = StringUtils.isEmpty(spacePerson.getOwnerSpace()) ? 0 : Double.parseDouble(spacePerson.getOwnerSpace());
         if (needSpacer > ownSpacer * 1073741824){
             throw new OtherException("个人可用空间不足，请扩容个人空间");
         }
@@ -142,13 +142,13 @@ public class StorageSpacePersonServiceImpl extends BaseServiceImpl<StorageSpaceP
     }
 
     @Override
-    public Long getPersonSpace(String userId, String storageType) {
+    public Double getPersonSpace(String userId, String storageType) {
         StorageSpace storageSpace = storageSpaceService.getByStorageType(storageType);
         if (null == storageSpace){
             throw new OtherException("存储空间不存在");
         }
         StorageSpacePerson byUserId = getByUserId(storageSpace.getId(),userId);
-        return Objects.isNull(byUserId) ? 0L : Opt.ofBlankAble(byUserId.getOwnerSpace()).map(Long::parseLong).orElse(0L);
+        return Objects.isNull(byUserId) ? 0L : Opt.ofBlankAble(byUserId.getOwnerSpace()).map(Double::parseDouble).orElse(0.0);
     }
 
     @Override
